requirements.txt
SQLAlchemy==1.3.23
requests==2.25.1
qrcode==6.1
Werkzeug==1.0.1
twilio==6.50.1
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
Flask_JWT_Extended==4.1.0
Flask_Cors==3.0.10
Flask_SocketIO==5.0.1
python_bcrypt==0.3.2
